requirements.txt
pprint==0.1
speedtest-cli==2.1.2
